Split crud.py into smaller modules
